패키지 구조 정리, Stream API 및 Cars 일급 컬렉션 적용
