+++ README.md
@@ -19,11 +19,16 @@
 - 전달받은 숫자가 4 이상이면 전진한다.
 - 전달받은 숫자가 3 이하이면 멈춘다.
 
-### RacingGame
+### Cars
 - 여러 대의 자동차를 관리한다.
 - 한 번의 라운드에서 모든 자동차를 전진 또는 멈추게 한다.
-- 가장 멀리 이동한 자동차를 우승자로 반환한다.
+- 가장 멀리 이동한 자동차의 위치를 기준으로 우승자를 판별한다.
 - 공동 우승자를 처리할 수 있다.
+
+### RacingGame
+- 자동차 경주 게임의 진행을 담당한다.
+- 한 번의 라운드를 진행한다.
+- 우승자 조회를 자동차 목록 객체에 위임한다.
 
 ### NumberGenerator
 - 숫자 생성 역할을 분리하기 위해 `NumberGenerator` 인터페이스를 사용했다.
@@ -42,6 +47,8 @@
 - 다수가 우승하는 경우를 테스트한다.
 
 ## 설계 의도
-`RacingGame` 내부에서 직접 랜덤값을 생성하면 테스트 결과를 예측하기 어렵다고 판단했습니다.  
-그래서 숫자 생성 역할을 `NumberGenerator`로 분리했습니다.  
+`NumberGenerator` 인터페이스를 도입하여 숫자 생성 역할을 분리했습니다.  
 이를 통해 실제 실행에서는 `RandomNumberGenerator`를 사용하고, 테스트에서는 `TestNumberGenerator`를 사용하여 원하는 시나리오를 검증할 수 있도록 했습니다.
+
+또한 `List<Car>`를 `Cars`로 감싸는 방식으로 자동차 목록에 대한 책임을 분리했습니다.  
+이를 통해 자동차 목록과 관련된 로직(전체 이동, 최대 위치 계산, 우승자 판별)을 `Cars` 내부에서 관리하고, `RacingGame`은 게임 진행에 집중하도록 구성했습니다.
