--- src/main/java/RacingGame.java
@@ -1,48 +1,0 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class RacingGame {
-
-    private final List<Car> cars;
-    private final NumberGenerator numberGenerator;
-
-    public RacingGame(List<Car> cars, NumberGenerator numberGenerator) {
-        this.cars = cars;
-        this.numberGenerator = numberGenerator;
-    }
-
-    public void playRound() {
-        for (Car car : cars) {
-            moveCar(car);
-        }
-    }
-
-    public List<Car> getWinners() {
-        int maxPosition = getMaxPosition();
-        List<Car> winners = new ArrayList<>();
-
-        for (Car car : cars) {
-            addWinner(winners, car, maxPosition);
-        }
-        return winners;
-    }
-
-    private void moveCar(Car car) {
-        car.move(numberGenerator.generate());
-    }
-
-    private int getMaxPosition() {
-        int maxPosition = 0;
-
-        for (Car car : cars) {
-            maxPosition = Math.max(maxPosition, car.getPosition());
-        }
-        return maxPosition;
-    }
-
-    private void addWinner(List<Car> winners, Car car, int maxPosition) {
-        if (car.getPosition() == maxPosition) {
-            winners.add(car);
-        }
-    }
-}
