+++ src/main/java/racingcar/domain/Cars.java
@@ -1,0 +1,35 @@
+package racingcar.domain;
+
+import racingcar.generator.NumberGenerator;
+
+import java.util.List;
+
+public class Cars {
+
+    private final List<Car> cars;
+
+    public Cars(List<Car> cars) {
+        this.cars = cars;
+    }
+
+    public void moveAll(NumberGenerator numberGenerator) {
+        for (Car car : cars) {
+            car.move(numberGenerator.generate());
+        }
+    }
+
+    public List<Car> getWinners() {
+        int maxPosition = getMaxPosition();
+
+        return cars.stream()
+                .filter(car -> car.getPosition() == maxPosition)
+                .toList();
+    }
+
+    private int getMaxPosition() {
+        return cars.stream()
+                .mapToInt(Car::getPosition)
+                .max()
+                .orElse(0);
+    }
+}
