+++ src/main/java/racingcar/generator/NumberGenerator.java
@@ -1,3 +1,5 @@
+package racingcar.generator;
+
 public interface NumberGenerator {
     int generate();
 }
