+++ src/main/java/racingcar/generator/RandomNumberGenerator.java
@@ -1,3 +1,5 @@
+package racingcar.generator;
+
 import java.util.Random;
 
 public class RandomNumberGenerator implements NumberGenerator {
