+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,3 +1,5 @@
+package racingcar.domain;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
